Organize code with components directory
