+++ src/client/src/api/stratigraphy.ts
@@ -41,6 +41,15 @@ export interface BaseLayer {
   isAutoCorrected?: boolean;
 }
 
+export interface DepthLayer {
+  id: string;
+  fromDepth: number;
+  toDepth: number;
+  hasFromDepthError?: boolean;
+  hasToDepthError?: boolean;
+  isAutoCorrected?: boolean;
+}
+
 interface Chronostratigraphy {
   id: number;
   stratigraphyId: number;
