+++ src/client/src/pages/detail/form/stratigraphy/components/addEmptyStratigraphyDialog/addEmptyStratigraphyDialog.tsx
@@ -1,11 +1,11 @@
 import { FC, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { Dialog, DialogActions, DialogContent, DialogTitle, Stack, TextField, Typography } from "@mui/material";
-import { ApiError } from "../../../../api/apiInterfaces.ts";
-import { Stratigraphy, useStratigraphyMutations } from "../../../../api/stratigraphy.ts";
-import { BoreholesButton, CancelButton } from "../../../../components/buttons/buttons.tsx";
-import { getFieldBorderColor } from "../../../../components/form/formUtils.ts";
-import { useApiErrorAlert } from "../../../../hooks/useShowAlertOnError.tsx";
+import { ApiError } from "../../../../../../api/apiInterfaces.ts";
+import { Stratigraphy, useStratigraphyMutations } from "../../../../../../api/stratigraphy.ts";
+import { BoreholesButton, CancelButton } from "../../../../../../components/buttons/buttons.tsx";
+import { getFieldBorderColor } from "../../../../../../components/form/formUtils.ts";
+import { useApiErrorAlert } from "../../../../../../hooks/useShowAlertOnError.tsx";
 
 interface AddEmptyStratigraphyDialogProps {
   open: boolean;
