+++ src/client/src/pages/detail/form/stratigraphy/components/lithologyTable/useLithologyTableState.ts
@@ -1,12 +1,11 @@
 import { useCallback, useMemo, useRef, useState } from "react";
 import { v4 as uuidv4 } from "uuid";
-import { BaseLayer } from "../../../../api/stratigraphy.ts";
-import { FaciesDescription } from "./faciesDescription.ts";
-import { LithologicalDescription } from "./lithologicalDescription.ts";
-import { Lithology } from "./lithology.ts";
+import { BaseLayer, DepthLayer } from "../../../../../../api/stratigraphy.ts";
+import { FaciesDescription } from "../../faciesDescription.ts";
+import { LithologicalDescription } from "../../lithologicalDescription.ts";
+import { Lithology } from "../../lithology.ts";
 import {
   createEmptyLithology,
-  DepthLayer,
   flagErrors,
   getInitialDepthLayers,
   mergeAdjacentDepths,
