+++ src/client/src/pages/detail/form/stratigraphy/stratigraphyPanel.tsx
@@ -18,9 +18,9 @@ import { useBoreholesNavigate } from "../../../../hooks/useBoreholesNavigate";
 import { useRequiredParams } from "../../../../hooks/useRequiredParams";
 import { formatDate } from "../../../../utils";
 import { EditStateContext } from "../../editStateContext";
-import { AddEmptyStratigraphyDialog } from "./addEmptyStratigraphyDialog.tsx";
-import { AddStratigraphyButton } from "./addStratigraphyButton";
 import ChronostratigraphyPanel from "./chronostratigraphy/chronostratigraphyPanel";
+import { AddEmptyStratigraphyDialog } from "./components/addEmptyStratigraphyDialog/addEmptyStratigraphyDialog.tsx";
+import { AddStratigraphyButton } from "./components/addStratigraphyButton.tsx";
 import { StratigraphyExtraction } from "./extraction/stratigraphyExtraction.tsx";
 import { LithologyPanel } from "./lithology/lithologyPanel.tsx";
 import LithostratigraphyPanel from "./lithostratigraphy/lithostratigraphyPanel";
